Implement skeletal IPA parsing
Prints the IPA entry from the nth line of the english IPA dictionary.

Currently use via:
`cabal --run -- --file NUM`
where NUM denotes the IPA entry you wish to access.

Next step is to have it do a lookup on a given word and return the
IPA for _that_ instead.