Create rudimentary typed stanza representation
As-is, only supports Fornyrðislag, and a rather restricted form of it
at that -- at least with regards to stanza length and alliterations.
The number of unstressed syllables is sadly completely unrestrained as
of yet, something I should definitely fix in the future.

Of much help in determining how best to restrict the form of the style
was the following resource:
http://digicoll.library.wisc.edu/Jonas/Prosody/Prosody-I.html

Could use significant improvements overall to improve type-level
strictness while preserving basic usability.