Improve line/halfline representation
Makes Line the GADT instead of HalfLine, within the hierarchy of
HalfLine ⊂ Line ⊂ Stanza.

It makes more sense to have the hidden type / existensial typeiness
happen at the Line level; after all, that's where the alliteration
actually matters (in that it links the two half-lines together
structurally/prosodically).

The following was the case with the prior version too, of course, but
just as a note: the fact that this representation makes the full type (and
thus the alliteration class) of the non-alliterating lift symbols is
intentional; their nature is rarely considered in the texts I've read on
the matter.

Only concern:
TODO: Somehow prevent non-allit sylls from allit'ing...